git-svn-id: https://39.108.144.9:8443/svn/puzhen@200 4ea25c84-54c4-e74c-a3e5-7ec0ea0d9549

+++ 2.0.0/source/application/store/controller/distributor/DistributorApply.php
@@ -27,7 +27,7 @@ class DistributorApply extends Controller
     public function edit($apply_id)
     {
         // 服务网点申请详情
-        $model = DistributorApplyModel::detail($apply_id);
+        $model = DistributorApplyModel::detail(['apply_id' => $apply_id]);
         if (!$this->request->isAjax()) {
             return $this->fetch('edit', compact('model'));
         }
@@ -46,12 +46,22 @@ class DistributorApply extends Controller
      */
     public function delete($apply_id)
     {
-        // 服务网点申请详情
-        $model = DistributorApplyModel::detail($apply_id);
-        if (!$model->delete()) {
-            return $this->renderError($model->getError() ?: '删除失败');
+        if(is_array($apply_id))
+        {
+            $model = new DistributorApplyModel;
+            $rst = $model->batchRemove($apply_id);
+            if (!$rst['status']) {
+                return $this->renderError($rst['message'] ?: '删除失败');
+            }
+            return $this->renderSuccess($rst['message'] ?: '删除成功');
+        }else{
+            // 服务网点申请详情
+            $model = DistributorApplyModel::detail(['apply_id' => $apply_id]);
+            if (!$model->delete()) {
+                return $this->renderError($model->getError() ?: '删除失败');
+            }
+            return $this->renderSuccess('删除成功');
         }
-        return $this->renderSuccess('删除成功');
     }
 
 }
